Add in Darwin functionality for Redis
Previously, the redis module had stripped the additional Darwin-specific
functionality out of the main redis class for refactoring. This change
re-introduces the code that was stripped into the redis::darwin class.
Functionally, the code is identical to how it existed previously. The main
redis rspec-puppet spec only introduces a check for the redis::darwin class -
its functionality will be checked within a later commit.