Managing {configdir,logdir,libdir} into redis class
Previously, the configdir, logdir, and libdir were being managed out of
the darwin class.  Because these are directories that are necessary to
be managed regardless of OS, they should be managed in the main redis class.

Currently there's a problem where Puppet is throwing an error that a File[undef]
is showing up with a duplicate definition.  We ran out of time and weren't able
to discover where this was coming from.

This starts us along the path for extracing Darwin and Debian-specific logic into
their own classes so the module can be more...modular.